feat: add authentication and project CRUD API routes with Prisma client setup

+++ src/app/api/projects/[projectId]/route.ts
@@ -1,8 +1,8 @@
 import { Prisma } from "@/generated/prisma/client";
 import { NextResponse, NextRequest } from "next/server";
-import { CreateProjectSchema } from "../../../../lib/validator/project";
-import { auth } from "../../../../lib/auth";
-import { prisma } from "../../../../lib/prisma";
+import { CreateProjectSchema } from "../../../../../lib/validator/project";
+import { auth } from "../../../../../lib/auth";
+import { prisma } from "../../../../../lib/prisma";
 import { headers } from "next/headers";
 
 // Create project
@@ -80,8 +80,11 @@ export async function GET(req: NextRequest) {
   }
 }
 
-// Delete projects
-export async function DELETE(req: NextRequest) {
+// update Project
+export async function PATCH(
+  req: NextRequest,
+  { params }: { params: Promise<{ projectId: string }> },
+) {
   try {
     const session = await auth.api.getSession({
       headers: req.headers,
@@ -91,18 +94,83 @@ export async function DELETE(req: NextRequest) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    const projects = await prisma.project.deleteMany({
+    const { projectId } = await params;
+
+    const body = await req.json();
+
+    const { name, domain } = body;
+
+    const existingProject = await prisma.project.findFirst({
       where: {
+        id: projectId,
         ownerId: session.user.id,
       },
     });
 
-    return NextResponse.json({ projects });
+    if (!existingProject) {
+      return NextResponse.json({ error: "Project not found" }, { status: 404 });
+    }
+
+    const Project = await prisma.project.update({
+      where: {
+        id: projectId,
+      },
+      data: {
+        ...(name !== undefined && { name }),
+        ...(domain !== undefined && { domain }),
+      },
+    });
+
+    return NextResponse.json({ Project }, { status: 200 });
   } catch (error) {
-    console.error("Failed to delete project:", error);
+    console.error("Failed to fetch projects:", error);
 
     return NextResponse.json(
-      { error: "Failed to delete project" },
+      { error: "Failed to fetch projects" },
+      { status: 500 },
+    );
+  }
+}
+
+// delete Project
+export async function DELETE(
+  req: NextRequest,
+  { params }: { params: Promise<{ projectId: string }> },
+) {
+  try {
+    const session = await auth.api.getSession({
+      headers: req.headers,
+    });
+
+    if (!session?.user) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    const { projectId } = await params;
+
+    const existingProject = await prisma.project.deleteMany({
+      where: {
+        id: projectId,
+        ownerId: session.user.id,
+      },
+    });
+
+    if (!existingProject) {
+      return NextResponse.json({ error: "Project not found" }, { status: 404 });
+    }
+
+    const Project = await prisma.project.delete({
+      where: {
+        id: projectId,
+      },
+    });
+
+    return NextResponse.json({ Project }, { status: 200 });
+  } catch (error) {
+    console.error("Failed to fetch projects:", error);
+
+    return NextResponse.json(
+      { error: "Failed to fetch projects" },
       { status: 500 },
     );
   }
